Move Plugin model to its own file
